+++ src/Make_vms.mms
@@ -318,6 +318,7 @@ SRC = \
 	bufwrite.c \
 	change.c \
 	charset.c \
+	cindent.c \
 	cmdexpand.c \
 	cmdhist.c \
 	crypt.c \
@@ -420,6 +421,7 @@ OBJ = \
 	bufwrite.obj \
 	change.obj \
 	charset.obj \
+	cindent.obj \
 	cmdexpand.obj \
 	cmdhist.obj \
 	crypt.obj \
@@ -697,6 +699,10 @@ change.obj : change.c vim.h [.auto]config.h feature.h os_unix.h \
  gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
  globals.h version.h
 charset.obj : charset.c vim.h [.auto]config.h feature.h os_unix.h \
+ ascii.h keymap.h term.h macros.h structs.h regexp.h \
+ gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
+ globals.h
+cindent.obj : cindent.c vim.h [.auto]config.h feature.h os_unix.h \
  ascii.h keymap.h term.h macros.h structs.h regexp.h \
  gui.h beval.h [.proto]gui_beval.pro option.h ex_cmds.h proto.h \
  globals.h
