+++ src/proto/ops.pro
@@ -12,7 +12,6 @@ int swapchar(int op_type, pos_T *pos);
 void op_insert(oparg_T *oap, long count1);
 int op_change(oparg_T *oap);
 void adjust_cursor_eol(void);
-int preprocs_left(void);
 char_u *skip_comment(char_u *line, int process, int include_space, int *is_comment);
 int do_join(long count, int insert_space, int save_undo, int use_formatoptions, int setmark);
 int fex_format(linenr_T lnum, long count, int c);
